+++ 3rd/nanovg/demos/agg_fill.c
@@ -13,7 +13,7 @@
 static void run_test(int32_t w, int32_t h, int32_t BPP, const char* filename) {
   int32_t size = w * h * BPP;
   uint8_t* data = (uint8_t*)malloc(size);
-  NVGcontext* vg = nvgCreateAGG(w, h, BPP == 2 ? NVG_TEXTURE_BGR565 : NVG_TEXTURE_RGBA, data);
+  NVGcontext* vg = nvgCreateAGG(w, h, w*BPP, BPP == 2 ? NVG_TEXTURE_BGR565 : NVG_TEXTURE_RGBA, data);
 
   memset(data, 0xff, size);